LayoutTests/platform/wpe/fast/speechsynthesis/speech-synthesis-speak-expected.txt

CONSOLE MESSAGE: TypeError: window.internals.enableMockSpeechSynthesizer is not a function. (In 'window.internals.enableMockSpeechSynthesizer()', 'window.internals.enableMockSpeechSynthesizer' is undefined)
FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
